Don't add keyboard-open to CSS if keyboard is already hidden
The 'keyboard-open' CSS class is added to body after a 400ms delay, but
nothing stops somebody from hiding the keyboard during this short period
of time, for example by tapping on empty space.

So we need to kill the timer responsible for adding this class when the
keyboard hides.

See this comment for some more details:

https://github.com/driftyco/ionic/issues/3041#issuecomment-276296886